final_project/requirements.txt
firecrawl-py
langchain
langchain-openai
langgraph
pydantic
python-dotenv
langchain_groq
langchain_community